Fixed all (I think) of the CodeQL issues.  A number of them were false positives or not worth fixing: like restructuring the methods of the ArrayOfNumbersSerDe to use the Visitor Pattern.  That is just too much work, generates a whole slew of subclasses which would probably be slower.
There may be a few CodeQL issues I didn't catch, which this push may
reveal. Then I'll deal with those.

+++ src/main/java/org/apache/datasketches/common/ArrayOfDoublesSerDe.java
@@ -66,7 +66,7 @@ public class ArrayOfDoublesSerDe extends ArrayOfItemsSerDe<Double> {
     Objects.requireNonNull(mem, "Memory must not be null");
     if (numItems <= 0) { return new Double[0]; }
     long offset = offsetBytes;
-    Util.checkBounds(offset, Double.BYTES * numItems, mem.getCapacity());
+    Util.checkBounds(offset, Double.BYTES * (long)numItems, mem.getCapacity());
     final Double[] array = new Double[numItems];
 
     for (int i = 0; i < numItems; i++) {
